infrastructure/functions/requirements.txt
google-cloud-run==0.10.2
google-api-core==2.15.0

